+++ code/ui/manager/src/components-new/Preview/Preview.tsx
